requirements.txt
streamlit 
pandas 
numpy 
qrcode==7.4.2 
pillow 
